src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/reference/ArrayReferenceShortAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.attribute.ArrayAttributeParser;
import org.openqa.selenium.WebElement;

public class ArrayReferenceShortAttributeParser extends ArrayAttributeParser {

    protected ArrayReferenceShortAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Short.parseShort(element.getAttribute(super.attribute));
    }

}
